chore(136): split tests/test_indexer.py into 6 files by feature
Split the 3,679-line test_indexer.py into:
- test_indexer_core.py (305 lines) — version, dialect, validation, checksums, structural diff, utils
- test_indexer_integration.py (769 lines) — full reindex cycle, MCP integration, phantom nodes, cross-file edges
- test_indexer_reindex.py (607 lines) — reindex_files variants (plain sql/dbt/sqlmesh), repo resolution
- test_indexer_columns.py (513 lines) — column persistence, batch insert, get_table_columns, null types
- test_indexer_schema.py (938 lines) — conventions/tags schema, sqlmesh batch + lineage, cross-repo columns
- test_indexer_dbt_schema.py (627 lines) — dbt ColumnDefResult persistence (schema.yml, sources)

All 98 tests still pass. No behavior changes.
